port permute
